Moved to queries directory + custom hook
