+++ src/components/apollo-client/queries/getExpenses.js
@@ -1,0 +1,22 @@
+import { gql, useQuery } from "@apollo/client";
+
+export const GET_EXPENSES = gql`
+query GetExpenses($email: String!) {
+  user(email: $email) {
+      currentExpenses {
+          amount
+          pctChange
+      }
+  }
+}`
+
+export const useGetExpenses = (email) => {
+  const { loading, error, data } = useQuery(GET_EXPENSES, {
+    variables: { email: email },
+  });
+  let totalExpensesData = null;
+  if (!loading && data) {
+    totalExpensesData = data?.user?.currentExpenses?.amount?.toFixed(2);
+  }
+  return { loading, error, totalExpensesData };
+};
