+++ src/components/apollo-client/queries/getIncomes.js
@@ -1,0 +1,22 @@
+import { gql, useQuery } from "@apollo/client";
+
+export const GET_INCOMES = gql`
+query GetIncomes($email: String!) {
+  user(email: $email) {
+      currentIncomes {
+          amount
+          pctChange
+      }
+  }
+}`
+
+export const useGetIncomes = (email) => {
+  const { loading, error, data } = useQuery(GET_INCOMES, {
+    variables: { email: email },
+  });
+  let totalIncomeData = null;
+  if (!loading && data) {
+    totalIncomeData = data?.user?.currentIncomes?.amount?.toFixed(2);
+  }
+  return { loading, error, totalIncomeData };
+};
