--- src/queries/getIncomes.js
@@ -1,11 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const GET_INCOMES = gql`
-query GetIncomes($email: String!) {
-  user(email: $email) {
-      currentIncomes {
-          amount
-          pctChange
-      }
-  }
-}`
